requirements.txt
gradio==5.49.0
fal-client>=0.5.0
huggingface-hub>=0.26.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

